general costs

+++ app/Services/ObjectGeneralCostService.php
@@ -7,14 +7,16 @@ use App\Models\Object\GeneralCost;
 
 class ObjectGeneralCostService
 {
-    public static function updateGeneralCost(BObject $object, float $amount): void
+    public static function updateGeneralCost(BObject $object, float $amount, bool $pinned = false, bool $needDelete = true): void
     {
-        GeneralCost::where('object_id', $object->id)->delete();
+        if ($needDelete) {
+            GeneralCost::where('object_id', $object->id)->delete();
+        }
 
         $generalCost = new GeneralCost();
         $generalCost->object_id = $object->id;
         $generalCost->amount = $amount;
-        $generalCost->is_pinned = false;
+        $generalCost->is_pinned = $pinned;
         $generalCost->save();
     }
 }
